+++ javascript/node/selenium-webdriver/test/firefox/firefox_test.js
@@ -18,6 +18,7 @@
 var path = require('path');
 
 var firefox = require('../../firefox'),
+    io = require('../../io'),
     test = require('../../lib/test'),
     assert = require('../../testing/assert');
 
@@ -119,6 +120,37 @@ test.suite(function(env) {
           return driver.isElementPresent({id: 'jetpack-sample-banner'});
         }, 3000);
       }
+    });
+
+    describe('profile management', function() {
+      var driver;
+
+      test.beforeEach(function() {
+        driver = null;
+      });
+
+      test.afterEach(function() {
+        if (driver) {
+          driver.quit();
+        }
+      });
+
+      test.it('deletes the temp profile on quit', function() {
+        driver = env.builder().build();
+
+        var profilePath = driver.call(function() {
+          var path = driver.profilePath_;
+          assert(io.exists(path)).isTrue();
+          return path;
+        });
+
+        return driver.quit().then(function() {
+          driver = null;
+          return profilePath;
+        }).then(function(path) {
+          assert(io.exists(path)).isFalse();
+        });
+      });
     });
   });
 }, {browsers: ['firefox']});